Convert Pedersen scripts to tests
